+++ Tests/Unit/Hooks/PageRenderer/SchemaMarkupInjectionTest.php
@@ -11,8 +11,9 @@ declare(strict_types=1);
 
 namespace Brotkrueml\Schema\Tests\Unit\Hooks\PageRenderer;
 
+use Brotkrueml\Schema\Adapter\ApplicationType;
+use Brotkrueml\Schema\Adapter\ExtensionAvailability;
 use Brotkrueml\Schema\Cache\PagesCacheService;
-use Brotkrueml\Schema\Event\ShouldEmbedMarkupEvent;
 use Brotkrueml\Schema\Extension;
 use Brotkrueml\Schema\Hooks\PageRenderer\SchemaMarkupInjection;
 use Brotkrueml\Schema\Manager\SchemaManager;
@@ -21,11 +22,10 @@ use Brotkrueml\Schema\Tests\Helper\SchemaCacheTrait;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\MockObject\Stub;
 use PHPUnit\Framework\TestCase;
-use TYPO3\CMS\Core\Cache\Frontend\FrontendInterface;
 use TYPO3\CMS\Core\Configuration\ExtensionConfiguration;
-use TYPO3\CMS\Core\EventDispatcher\EventDispatcher;
-use TYPO3\CMS\Core\Http\ApplicationType;
+use TYPO3\CMS\Core\Package\PackageManager;
 use TYPO3\CMS\Core\Page\PageRenderer;
+use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 
@@ -56,14 +56,14 @@ class SchemaMarkupInjectionTest extends TestCase
     private $pagesCacheServiceMock;
 
     /**
-     * @var Stub|EventDispatcher
-     */
-    private $eventDispatcherStub;
-
-    /**
      * @var Stub|ApplicationType
      */
     private $applicationTypeStub;
+
+    /**
+     * @var Stub|ExtensionAvailability
+     */
+    private $extensionAvailabilityStub;
 
     protected function setUp(): void
     {
@@ -72,24 +72,17 @@ class SchemaMarkupInjectionTest extends TestCase
         $this->controllerMock->page = ['no_index' => 0, 'uid' => 42];
 
         $this->extensionConfigurationMock = $this->createMock(ExtensionConfiguration::class);
-
         $this->pagesCacheServiceMock = $this->createMock(PagesCacheService::class);
-
-        $this->eventDispatcherStub = $this->createStub(EventDispatcher::class);
-        $this->eventDispatcherStub
-            ->method('dispatch')
-            ->with(self::anything())
-            ->willReturn(new ShouldEmbedMarkupEvent([], true));
-
         $this->applicationTypeStub = $this->createStub(ApplicationType::class);
+        $this->extensionAvailabilityStub = $this->createStub(ExtensionAvailability::class);
 
         $this->subject = new SchemaMarkupInjection(
             $this->controllerMock,
             $this->extensionConfigurationMock,
             GeneralUtility::makeInstance(SchemaManager::class),
             $this->pagesCacheServiceMock,
-            $this->eventDispatcherStub,
-            $this->applicationTypeStub
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
         );
 
         $this->pageRendererMock = $this->createMock(PageRenderer::class);
@@ -178,13 +171,17 @@ class SchemaMarkupInjectionTest extends TestCase
             ->method('isBackend')
             ->willReturn(false);
 
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(false);
+
         $subject = new SchemaMarkupInjection(
             $this->controllerMock,
             $this->extensionConfigurationMock,
             $schemaManager,
             $this->pagesCacheServiceMock,
-            $this->eventDispatcherStub,
-            $this->applicationTypeStub
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
         );
 
         $params = [];
@@ -221,13 +218,17 @@ class SchemaMarkupInjectionTest extends TestCase
             ->method('isBackend')
             ->willReturn(false);
 
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(false);
+
         $subject = new SchemaMarkupInjection(
             $this->controllerMock,
             $this->extensionConfigurationMock,
             GeneralUtility::makeInstance(SchemaManager::class),
             $this->pagesCacheServiceMock,
-            $this->eventDispatcherStub,
-            $this->applicationTypeStub
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
         );
 
         $params = [];
@@ -252,13 +253,17 @@ class SchemaMarkupInjectionTest extends TestCase
             ->method('isBackend')
             ->willReturn(false);
 
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(false);
+
         $subject = new SchemaMarkupInjection(
             $controllerMock,
             $this->extensionConfigurationMock,
             null,
             $this->pagesCacheServiceMock,
-            $this->eventDispatcherStub,
-            $this->applicationTypeStub
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
         );
 
         $schemaManager = GeneralUtility::makeInstance(SchemaManager::class);
@@ -279,7 +284,7 @@ class SchemaMarkupInjectionTest extends TestCase
     /**
      * @test
      */
-    public function whenCacheIDefinedItIsUsedToGetMarkup(): void
+    public function whenCacheIsDefinedItIsUsedToGetMarkup(): void
     {
         $this->pagesCacheServiceMock
             ->expects(self::once())
@@ -295,13 +300,17 @@ class SchemaMarkupInjectionTest extends TestCase
             ->method('isBackend')
             ->willReturn(false);
 
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(false);
+
         $subject = new SchemaMarkupInjection(
             $this->controllerMock,
             $this->extensionConfigurationMock,
             null,
             $this->pagesCacheServiceMock,
-            $this->eventDispatcherStub,
-            $this->applicationTypeStub
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
         );
 
         $params = [];
@@ -333,13 +342,17 @@ class SchemaMarkupInjectionTest extends TestCase
             ->method('isBackend')
             ->willReturn(false);
 
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(false);
+
         $subject = new SchemaMarkupInjection(
             $this->controllerMock,
             $this->extensionConfigurationMock,
             $schemaManager,
             $this->pagesCacheServiceMock,
-            $this->eventDispatcherStub,
-            $this->applicationTypeStub
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
         );
 
         $params = [];
@@ -349,26 +362,222 @@ class SchemaMarkupInjectionTest extends TestCase
     /**
      * @test
      */
-    public function eventDispatcherForShouldEmbedMarkupEventReturnsFalseThenNoMarkupIsEmbedded(): void
+    public function whenPageShouldBeIndexedThenMarkupIsEmbedded(): void
     {
         /** @var SchemaManager $schemaManager */
         $schemaManager = GeneralUtility::makeInstance(SchemaManager::class);
         $schemaManager->addType((new Thing())->setProperty('name', 'some name'));
 
-        $cacheMock = $this->createMock(FrontendInterface::class);
-        $cacheMock
-            ->expects(self::never())
-            ->method('get')
-            ->willReturn(false);
-
-        $eventDispatcherStub = $this->createStub(EventDispatcher::class);
-        $eventDispatcherStub
-            ->method('dispatch')
-            ->with(self::anything())
-            ->willReturn(new ShouldEmbedMarkupEvent([], false));
+        $this->pagesCacheServiceMock
+            ->expects(self::once())
+            ->method('getMarkupFromCache')
+            ->willReturn(null);
+        $this->pagesCacheServiceMock
+            ->expects(self::once())
+            ->method('storeMarkupInCache')
+            ->with(\sprintf(
+                Extension::JSONLD_TEMPLATE,
+                '{"@context":"https://schema.org/","@type":"Thing","name":"some name"}'
+            ));
 
         $this->applicationTypeStub
             ->method('isBackend')
+            ->willReturn(false);
+
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(true);
+
+        $subject = new SchemaMarkupInjection(
+            $this->controllerMock,
+            $this->extensionConfigurationMock,
+            $schemaManager,
+            $this->pagesCacheServiceMock,
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
+        );
+
+        $params = [];
+        $subject->execute($params, $this->pageRendererMock);
+    }
+
+    /**
+     * @test
+     */
+    public function whenPageShouldNotBeIndexedAndConfigurationOptionIsNotDefinedThenMarkupIsEmbedded(): void
+    {
+        /** @var SchemaManager $schemaManager */
+        $schemaManager = GeneralUtility::makeInstance(SchemaManager::class);
+        $schemaManager->addType((new Thing())->setProperty('name', 'some name'));
+
+        $this->controllerMock->page = ['no_index' => 1, 'uid' => 42];
+
+        $this->extensionConfigurationMock
+            ->expects(self::once())
+            ->method('get')
+            ->with('schema')
+            ->willReturn([]);
+
+        $this->pagesCacheServiceMock
+            ->expects(self::once())
+            ->method('getMarkupFromCache')
+            ->willReturn(null);
+        $this->pagesCacheServiceMock
+            ->expects(self::once())
+            ->method('storeMarkupInCache')
+            ->with(\sprintf(
+                Extension::JSONLD_TEMPLATE,
+                '{"@context":"https://schema.org/","@type":"Thing","name":"some name"}'
+            ));
+
+        $this->applicationTypeStub
+            ->method('isBackend')
+            ->willReturn(false);
+
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(true);
+
+        $subject = new SchemaMarkupInjection(
+            $this->controllerMock,
+            $this->extensionConfigurationMock,
+            $schemaManager,
+            $this->pagesCacheServiceMock,
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
+        );
+
+        $params = [];
+        $subject->execute($params, $this->pageRendererMock);
+    }
+
+    /**
+     * @test
+     */
+    public function whenPageShouldNotBeIndexedAndConfigurationOptionIsActivatedThenMarkupIsEmbedded(): void
+    {
+        /** @var SchemaManager $schemaManager */
+        $schemaManager = GeneralUtility::makeInstance(SchemaManager::class);
+        $schemaManager->addType((new Thing())->setProperty('name', 'some name'));
+
+        $this->controllerMock->page = ['no_index' => 1, 'uid' => 42];
+
+        $this->extensionConfigurationMock
+            ->expects(self::once())
+            ->method('get')
+            ->with('schema')
+            ->willReturn(['embedMarkupOnNoindexPages' => '1']);
+
+        $this->pagesCacheServiceMock
+            ->expects(self::once())
+            ->method('getMarkupFromCache')
+            ->willReturn(null);
+        $this->pagesCacheServiceMock
+            ->expects(self::once())
+            ->method('storeMarkupInCache')
+            ->with(\sprintf(
+                Extension::JSONLD_TEMPLATE,
+                '{"@context":"https://schema.org/","@type":"Thing","name":"some name"}'
+            ));
+
+        $this->applicationTypeStub
+            ->method('isBackend')
+            ->willReturn(false);
+
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(true);
+
+        $subject = new SchemaMarkupInjection(
+            $this->controllerMock,
+            $this->extensionConfigurationMock,
+            $schemaManager,
+            $this->pagesCacheServiceMock,
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
+        );
+
+        $params = [];
+        $subject->execute($params, $this->pageRendererMock);
+    }
+
+    /**
+     * @test
+     */
+    public function whenPageShouldNotBeIndexedAndConfigurationOptionIsDeactivatedThenMarkupIsNotEmbedded(): void
+    {
+        /** @var SchemaManager $schemaManager */
+        $schemaManager = GeneralUtility::makeInstance(SchemaManager::class);
+        $schemaManager->addType((new Thing())->setProperty('name', 'some name'));
+
+        $this->controllerMock->page = ['no_index' => 1, 'uid' => 42];
+
+        $this->extensionConfigurationMock
+            ->expects(self::once())
+            ->method('get')
+            ->with('schema')
+            ->willReturn(['embedMarkupOnNoindexPages' => '0']);
+
+        $this->pagesCacheServiceMock
+            ->expects(self::never())
+            ->method('getMarkupFromCache');
+
+        $this->applicationTypeStub
+            ->method('isBackend')
+            ->willReturn(false);
+
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
+            ->willReturn(true);
+
+        $subject = new SchemaMarkupInjection(
+            $this->controllerMock,
+            $this->extensionConfigurationMock,
+            $schemaManager,
+            $this->pagesCacheServiceMock,
+            $this->applicationTypeStub,
+            $this->extensionAvailabilityStub
+        );
+
+        $params = [];
+        $subject->execute($params, $this->pageRendererMock);
+    }
+
+    /**
+     * @test
+     */
+    public function whenSeoExtensionIsNotLoadedMarkupIsAlwaysEmbedded(): void
+    {
+        /** @var SchemaManager $schemaManager */
+        $schemaManager = GeneralUtility::makeInstance(SchemaManager::class);
+        $schemaManager->addType((new Thing())->setProperty('name', 'some name'));
+
+        $this->controllerMock->page = ['no_index' => 1, 'uid' => 42];
+
+        $this->extensionConfigurationMock
+            ->expects(self::once())
+            ->method('get')
+            ->with('schema')
+            ->willReturn(['embedMarkupOnNoindexPages' => '0']);
+
+        $this->pagesCacheServiceMock
+            ->expects(self::once())
+            ->method('getMarkupFromCache')
+            ->willReturn(null);
+        $this->pagesCacheServiceMock
+            ->expects(self::once())
+            ->method('storeMarkupInCache')
+            ->with(\sprintf(
+                Extension::JSONLD_TEMPLATE,
+                '{"@context":"https://schema.org/","@type":"Thing","name":"some name"}'
+            ));
+
+        $this->applicationTypeStub
+            ->method('isBackend')
+            ->willReturn(false);
+
+        $this->extensionAvailabilityStub
+            ->method('isSeoAvailable')
             ->willReturn(false);
 
         $subject = new SchemaMarkupInjection(
@@ -376,9 +585,15 @@ class SchemaMarkupInjectionTest extends TestCase
             $this->extensionConfigurationMock,
             $schemaManager,
             $this->pagesCacheServiceMock,
-            $eventDispatcherStub,
             $this->applicationTypeStub
         );
+
+        $packageManagerStub = $this->createStub(PackageManager::class);
+        $packageManagerStub
+            ->method('isPackageActive')
+            ->with('seo')
+            ->willReturn(false);
+        ExtensionManagementUtility::setPackageManager($packageManagerStub);
 
         $params = [];
         $subject->execute($params, $this->pageRendererMock);